Move ability manager
