+++ apps/web/src/app/[locale]/(space)/layout.tsx
@@ -4,11 +4,11 @@ import Link from "next/link";
 import { redirect } from "next/navigation";
 import { loadCurrentUserSpace } from "@/auth/data";
 import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
-import { AbilityProvider } from "@/features/ability-manager/client";
 import { LicenseLimitWarning } from "@/features/licensing/components/license-limit-warning";
 import { CommandMenu } from "@/features/navigation/command-menu";
 import { isUserOnboarded } from "@/features/setup/utils";
 import { TimezoneProvider } from "@/features/timezone/client/context";
+import { AbilityProvider } from "@/lib/ability-manager/client";
 import { SpaceSidebar } from "./components/sidebar/space-sidebar";
 import { SpaceSidebarProvider } from "./components/sidebar/space-sidebar-provider";
 import { TopBar, TopBarLeft, TopBarRight } from "./components/top-bar";
