+++ apps/web/src/features/spaces/data.ts
@@ -6,9 +6,9 @@ import type {
 import { prisma } from "@rallly/database";
 import { cache } from "react";
 import { loadCurrentUser, loadCurrentUserSpace } from "@/auth/data";
-import { defineAbilityFor } from "@/features/ability-manager";
 import type { SpaceMemberRole } from "@/features/spaces/schema";
 import { fromDBRole, toDBRole } from "@/features/spaces/utils";
+import { defineAbilityFor } from "@/lib/ability-manager";
 import { AppError } from "@/lib/errors";
 import { isSelfHosted } from "@/utils/constants";
